requirements.txt
crewai==0.140.0
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
python-dotenv
requests
beautifulsoup4
pandas
numpy
matplotlib
seaborn
plotly
openai
anthropic
google-generativeai
tavily-python
duckduckgo-search
wikipedia
arxiv
scholarly
pymongo
json5
markdown
jinja2
nltk 